yacht-cloud/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "yacht-cloud"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn>=0.35.0",
  "pydantic>=2.8.0",
  "email-validator>=2.2.0",
  "python-dotenv>=1.0.1",
  "PyJWT>=2.9.0",
  "passlib>=1.7.4",
  "SQLAlchemy>=2.0.30",
  "psycopg[binary]>=3.2.1",
  "stripe>=10.10.0",
  "PyYAML>=6.0.1"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
